pyproject.toml
[tool.poetry]
name = "fconline-coach-agent"
version = "0.1.0"
description = ""
authors = ["hskimim <hskimim8855@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.13.3"
numpy = "^1.26.4"
pandas = "^2.2.1"
gradio = "^4.20.0"
python-dotenv = "^1.0.1"
llmlingua = "^0.1.6"
llama-index = "^0.10.17"
langchain = "^0.1.11"
langchain-community = "^0.0.27"
langchain-openai = "^0.0.8"
faiss-cpu = "^1.8.0"
accelerate = "^0.27.2"
torchaudio = "^2.2.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
